# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "MovePreferenceResponse": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "move",
            "title": "Kind",
            "type": "string"
          },
          "uci": {
            "title": "Uci",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "uci"
        ],
        "title": "MovePreferenceResponse",
        "type": "object"
      },
...
      "NoPreferenceResponse": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "no_preference",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "title": "NoPreferenceResponse",
        "type": "object"
      },
      "OpeningCatalogueErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "invalid_filter",
              "openings_unavailable",
              "unexpected_failure"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "OpeningCatalogueErrorResponse",
        "type": "object"
      },
      "OpeningCatalogueItemResponse": {
        "additionalProperties": false,
        "properties": {
          "eco": {
            "title": "Eco",
            "type": "string"
          },
          "games_deepest": {
            "title": "Games Deepest",
            "type": "integer"
          },
          "games_reached": {
            "title": "Games Reached",
            "type": "integer"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "route_count": {
            "title": "Route Count",
            "type": "integer"
          }
        },
        "required": [
          "key",
          "eco",
          "name",
          "route_count",
          "games_reached",
          "games_deepest"
        ],
        "title": "OpeningCatalogueItemResponse",
        "type": "object"
      },
      "OpeningCatalogueResponse": {
        "additionalProperties": false,
        "properties": {
          "has_next": {
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OpeningCatalogueItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "total_pages",
          "has_next"
        ],
        "title": "OpeningCatalogueResponse",
        "type": "object"
      },
      "OpeningDetailErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "invalid_opening_key",
              "opening_not_found",
              "openings_unavailable",
              "unexpected_failure"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "OpeningDetailErrorResponse",
        "type": "object"
      },
...
      "PositionInsightAnalysisResponse": {
        "additionalProperties": false,
        "properties": {
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PositionInsightResultResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "enum": [
              "not_requested",
              "queued",
              "running",
              "ready"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "result"
        ],
        "title": "PositionInsightAnalysisResponse",
        "type": "object"
      },
      "PositionInsightErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "invalid_fen",
              "invalid_trainer_color",
              "invalid_as_of",
              "position_insight_unavailable",
              "unexpected_failure"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "PositionInsightErrorResponse",
        "type": "object"
      },
      "PositionInsightExperienceResponse": {
        "additionalProperties": false,
        "properties": {
          "distinct_game_count": {
            "minimum": 0.0,
            "title": "Distinct Game Count",
            "type": "integer"
          },
          "occurrence_count": {
            "minimum": 0.0,
            "title": "Occurrence Count",
            "type": "integer"
          }
        },
        "required": [
          "distinct_game_count",
          "occurrence_count"
        ],
        "title": "PositionInsightExperienceResponse",
        "type": "object"
      },
      "PositionInsightLineResponse": {
        "additionalProperties": false,
        "properties": {
          "depth": {
            "minimum": 0.0,
            "title": "Depth",
            "type": "integer"
          },
          "pv_uci": {
            "items": {
              "type": "string"
            },
            "title": "Pv Uci",
            "type": "array"
          },
          "rank": {
            "minimum": 1.0,
            "title": "Rank",
            "type": "integer"
          },
          "score_kind": {
            "enum": [
              "cp",
              "mate"
            ],
            "title": "Score Kind",
            "type": "string"
          },
          "score_value": {
            "title": "Score Value",
            "type": "integer"
          },
          "wdl_draws": {
            "minimum": 0.0,
            "title": "Wdl Draws",
            "type": "integer"
          },
          "wdl_losses": {
            "minimum": 0.0,
            "title": "Wdl Losses",
            "type": "integer"
          },
          "wdl_wins": {
            "minimum": 0.0,
            "title": "Wdl Wins",
            "type": "integer"
          }
        },
        "required": [
          "rank",
          "score_kind",
          "score_value",
          "wdl_wins",
          "wdl_draws",
          "wdl_losses",
          "pv_uci",
          "depth"
        ],
        "title": "PositionInsightLineResponse",
        "type": "object"
      },
      "PositionInsightObservedMoveResponse": {
        "additionalProperties": false,
        "properties": {
          "distinct_game_count": {
            "minimum": 0.0,
            "title": "Distinct Game Count",
            "type": "integer"
          },
          "move_uci": {
            "title": "Move Uci",
            "type": "string"
          },
          "occurrence_count": {
            "minimum": 0.0,
            "title": "Occurrence Count",
            "type": "integer"
          }
        },
        "required": [
          "move_uci",
          "distinct_game_count",
          "occurrence_count"
        ],
        "title": "PositionInsightObservedMoveResponse",
        "type": "object"
      },
      "PositionInsightOpeningResponse": {
        "additionalProperties": false,
        "properties": {
          "eco": {
            "title": "Eco",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "match": {
            "enum": [
              "route",
              "transposition"
            ],
            "title": "Match",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ply": {
            "minimum": 0.0,
            "title": "Ply",
            "type": "integer"
          }
        },
        "required": [
          "key",
          "eco",
          "name",
          "ply",
          "match"
        ],
        "title": "PositionInsightOpeningResponse",
        "type": "object"
      },
      "PositionInsightResponse": {
        "additionalProperties": false,
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/PositionInsightAnalysisResponse"
          },
          "as_of": {
            "title": "As Of",
            "type": "string"
          },
          "experience": {
            "$ref": "#/components/schemas/PositionInsightExperienceResponse"
          },
          "fen": {
            "title": "Fen",
            "type": "string"
          },
          "observed_moves": {
            "items": {
              "$ref": "#/components/schemas/PositionInsightObservedMoveResponse"
            },
            "title": "Observed Moves",
            "type": "array"
          },
          "opening": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PositionInsightOpeningResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "preference": {
            "discriminator": {
              "mapping": {
                "move": "#/components/schemas/MovePreferenceResponse",
                "no_preference": "#/components/schemas/NoPreferenceResponse",
                "unconfigured": "#/components/schemas/UnconfiguredPreferenceResponse"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/MovePreferenceResponse"
              },
              {
                "$ref": "#/components/schemas/NoPreferenceResponse"
              },
              {
                "$ref": "#/components/schemas/UnconfiguredPreferenceResponse"
              }
            ],
            "title": "Preference"
          },
          "trainer_color": {
            "enum": [
              "white",
              "black"
            ],
            "title": "Trainer Color",
            "type": "string"
          }
        },
        "required": [
          "fen",
          "trainer_color",
          "as_of",
          "opening",
          "experience",
          "observed_moves",
          "analysis",
          "preference"
        ],
        "title": "PositionInsightResponse",
        "type": "object"
      },
      "PositionInsightResultResponse": {
        "additionalProperties": false,
        "properties": {
          "configuration_version": {
            "minimum": 1.0,
            "title": "Configuration Version",
            "type": "integer"
          },
          "engine_name": {
            "title": "Engine Name",
            "type": "string"
          },
          "engine_version": {
            "title": "Engine Version",
            "type": "string"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/PositionInsightLineResponse"
            },
            "title": "Lines",
            "type": "array"
          },
          "quality": {
            "enum": [
              "browser",
              "tool"
            ],
            "title": "Quality",
            "type": "string"
          },
          "settings": {
            "title": "Settings",
            "type": "object"
          },
          "terminal_kind": {
            "anyOf": [
              {
                "enum": [
                  "checkmate",
                  "stalemate",
                  "insufficient_material"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal Kind"
          }
        },
        "required": [
          "quality",
          "configuration_version",
          "settings",
          "engine_name",
          "engine_version",
          "terminal_kind",
          "lines"
        ],
        "title": "PositionInsightResultResponse",
        "type": "object"
      },
...
      "UnconfiguredPreferenceResponse": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "unconfigured",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "title": "UnconfiguredPreferenceResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Chess Move Trainer",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/openings": {
      "get": {
        "operationId": "getOpenings",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "eco_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Eco From"
            }
          },
          {
            "in": "query",
            "name": "eco_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Eco To"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "eco_asc",
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpeningCatalogueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpeningCatalogueErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpeningCatalogueErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpeningCatalogueErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Openings",
        "tags": [
          "openings"
        ]
      }
    },
    "/api/openings/{opening_key}": {
      "get": {
        "operationId": "getOpeningByKey",
        "parameters": [
          {
            "in": "path",
            "name": "opening_key",
            "required": true,
            "schema": {
              "title": "Opening Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpeningCatalogueItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpeningDetailErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpeningDetailErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpeningDetailErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpeningDetailErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Opening By Key",
        "tags": [
          "openings"
        ]
      }
    },
    "/api/positions/insight": {
      "get": {
        "operationId": "getPositionInsight",
        "parameters": [
          {
            "in": "query",
            "name": "fen",
            "required": true,
            "schema": {
              "title": "Fen",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "trainer_color",
            "required": true,
            "schema": {
              "title": "Trainer Color",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": true,
            "schema": {
              "title": "As Of",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionInsightResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionInsightErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionInsightErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionInsightErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Position Insight",
        "tags": [
          "position-insight"
        ]
      }
    },
...
  }
}